requirements.txt
git+git://github.com/Clarifai/Clarifai_py.git
Pillow
requests
MetaMindApi
